VMX: only check the PB VM EC2 if EC1 is ok
We should only check the Secondary Processor-Based VM-Execution controls
if the bit is present in the primary controls.  It's one of our
set-to-one bits.

If we do the secondary checks without that magic bit set, we'll GPF.  We
don't have rdmsr_safe() or anything like that either.

Ultimately, once 'ok' is false, we're going to fail anyway.  It's just a
question of how much info we get.