Create ros/vmm/vmx.h, prepare to move lots of code to user mode from kernel driver.
We need these constants and maybe these structs in user mode, so we can
handle more vmexits. Now that this is done, we can start to move
code out of the kernel driver.

Longer term, we probably even want the vmcs registers set up in user mode
so we can deal with the inevitable undocumented/incorrect hardware without
having to respin the kernel.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>