Start putting together ACPI for guests.
This sucks, but has to happen if we want to let guests use
x2apic without lots of mods to the kernel.

But I still hate acpi.

This needs to get all library-ized once I learn how to make it work.

Worst case is we end up writing ACPI code and compiling it with iasl.

I'd really rather avoid that.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>