Clean up awkward code sequences in ether driver.
Rework two awkward code sequences in the ethernet driver to
make the logic easier to follow.

In the first case, put explicit freeb() and early returns
into etheroq in the cases where we are looping a frame back
up the stack.

In the second, change the loop in etheroq to use 'guard'
conditions and early continues to make the logic linear.