Remove the special casing of #pipe in lseek
If we want to disallow seeks on a per-device basis, then we need a device
op (probably use chan_ctl).  None of this crappy special casing.
Incidentally, we'd probably want to do it for #ip too, and who knows what
is running on the other end of a server (or across a mount point!).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>