Add fdmux support.
fdmux acts kind of like ptys but is a bit more general purpose
and legacy free. There's lots missing here. But I think it's a good
direction.

You can do this:
fdmux
and it lands you in an sh. You can then run, e.g., cat,
hit ^C, and I suspect that the fdmux driver is sending a ^C.

The ctl file has several commands, including
d
for debug.

Lots of room for improvement here, consider this a 'checkpoint commit'

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>