dev/pipe.c: grow the max pipe size to 32MB
this is a hack. It gets scp going. The problem is that the pipe writer
stalls when the pipe is full but the reader does not wake up.

I don't know akaros scheduling well enough to fix this, but 32K pipes? So 1990, when
machines had 64M, not 64G.

I think 32M is a reasonable upper bound. But as soon as you try to copy
a file that is 32768*1024 + 1 bytes long, scp will stall again.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>